Add omxmdeiscaler element
The omxmdeiscaler element was added. Radeus Lab added the
two pads and the caps setting. Also the second output port
configuration. Ridgerun added the subsampling property to the
deiscaler configuration. The most changes done in order to get the
plugin to work were added to the videofilter class.
TODO: Improve the handling of the caps negotiation in the srcpads